+++ tests/test_data/test_monitor_data.py
@@ -23,6 +23,7 @@ from .test_data_arrays import FIELD_MONITOR, FIELD_TIME_MONITOR, MODE_MONITOR_WI
 from .test_data_arrays import MODE_MONITOR, PERMITTIVITY_MONITOR, FLUX_MONITOR, FLUX_TIME_MONITOR
 from .test_data_arrays import FIELD_MONITOR_2D, FIELD_TIME_MONITOR_2D
 from .test_data_arrays import DIFFRACTION_MONITOR, SIM_SYM, SIM
+from .test_data_arrays import FS
 from ..utils import assert_log_level
 from ..utils import log_capture  # noqa: F401
 
@@ -361,7 +362,7 @@ def test_empty_array():
         symmetry=SIM.symmetry,
         symmetry_center=SIM.center,
         grid_expanded=SIM.discretize_monitor(monitor),
-        **fields
+        **fields,
     )
 
 
@@ -375,7 +376,7 @@ def _test_empty_list():
         symmetry=SIM.symmetry,
         symmetry_center=SIM.center,
         grid_expanded=SIM.discretize_monitor(monitor),
-        **fields
+        **fields,
     )
 
 
@@ -389,7 +390,7 @@ def _test_empty_tuple():
         symmetry=SIM.symmetry,
         symmetry_center=SIM.center,
         grid_expanded=SIM.discretize_monitor(monitor),
-        **fields
+        **fields,
     )
 
 
@@ -402,7 +403,7 @@ def test_empty_io(tmp_path):
         symmetry=SIM.symmetry,
         symmetry_center=SIM.center,
         grid_expanded=SIM.discretize_monitor(monitor),
-        **fields
+        **fields,
     )
     field_data.to_file(str(tmp_path / "field_data.hdf5"))
     field_data = td.FieldTimeData.from_file(str(tmp_path / "field_data.hdf5"))
@@ -431,7 +432,7 @@ def test_field_data_symmetry_present():
             monitor=monitor,
             symmetry=(1, -1, 0),
             grid_expanded=SIM.discretize_monitor(monitor),
-            **fields
+            **fields,
         )
 
     # fails if symmetry specified but missing etended grid
@@ -530,11 +531,11 @@ def test_outer_dot():
     dot = mode_data.outer_dot(mode_data)
     assert "mode_index_0" in dot.coords and "mode_index_1" in dot.coords
     dot = field_data.outer_dot(mode_data)
-    assert not "mode_index_0" in dot.coords and "mode_index_1" in dot.coords
+    assert "mode_index_0" not in dot.coords and "mode_index_1" in dot.coords
     dot = mode_data.outer_dot(field_data)
-    assert "mode_index_0" in dot.coords and not "mode_index_1" in dot.coords
+    assert "mode_index_0" in dot.coords and "mode_index_1" not in dot.coords
     dot = field_data.outer_dot(field_data)
-    assert not "mode_index_0" in dot.coords and not "mode_index_1" in dot.coords
+    assert "mode_index_0" not in dot.coords and  "mode_index_1" not in dot.coords
 
     # test that only common freqs are kept
     inds1 = [0, 1, 3]
@@ -581,4 +582,12 @@ def test_no_nans():
         **{key: eps_nan for key in ["eps_xx", "eps_yy", "eps_zz"]}
     )
     with pytest.raises(pydantic.ValidationError):
-        cm = td.CustomMedium(eps_dataset=eps_dataset_nan)
+        _ = td.CustomMedium(eps_dataset=eps_dataset_nan)
+
+
+def test_zbf_output(tmp_path):
+    field_data = make_field_data_2d()
+    field_data.to_zbf(str(tmp_path / "field_data.zbf"), 1.0, FS[len(FS) // 2])
+
+    mode_data = make_mode_solver_data()
+    mode_data.to_zbf(str(tmp_path / "mode_data.zbf"), 1.0, FS[len(FS) // 2])